Refactor imports and split CHRR tests
