+++ src/cobra/sampling/hopsy.py
@@ -6,213 +6,205 @@ from typing import TYPE_CHECKING, Optional
 if TYPE_CHECKING:
     from cobra import Model
 
+import hopsy
+import numpy as np
+import pandas as pd
+
 from ..core.configuration import Configuration
 from .hr_sampler import HRSampler
 
 
-__all__ = (
-    "HopsySampler",
-    "hopsy_is_available",
-)
+__all__ = ("HopsySampler",)
 
 configuration = Configuration()
 
-try:
-    import hopsy
-    import numpy as np
-    import pandas as pd
 
-    hopsy_is_available = True
+# for the time being while the updated hopsy version isn't released
+def back_transform(problem, samples):
+    """
+    Transform samples from the sampling space to the original parameter space.
 
-    # for the time being while the updated hopsy version isn't released
-    def back_transform(problem, samples):
-        """
-        Transform samples from the sampling space to the original parameter space.
+    Parameters
+    ----------
+    problem : hopsy.Problem
+      A hopsy.Problem with populated `problem.transformation` and `problem.shift` of
+      shape `(m, d)` and `(d, )`, respectively. If not populated, the identity
+      operation is performed.
+    samples : np.array
+      Parameter samples of shape `(..., d)` which are to be transformed.
+
+    Returns
+    -------
+    np.array
+      Transformed samples of shape `(..., m)`
+    """
+    S, h = problem.transformation, problem.shift
+    transformed = samples @ S.T if S is not None else samples
+    shifted = transformed + h if h is not None else transformed
+    return shifted
+
+
+class HopsySampler(HRSampler):
+    """
+    Generic sampling wrapper for hopsy samplers.
+
+    Parameters
+    ----------
+    model : cobra.Model
+        The cobra model from which to generate samples.
+    sampler : hopsy.Proposal, optional
+        The hopsy sampler for sampling the polytope
+        (default hopsy.UniformCoordinateHitAndRunProposal).
+    processes: int, optional
+        The number of processes used during sampling
+        (default cobra.Configuration.processes).
+    thinning : int, optional
+        The thinning factor of the generated sampling chain. A thinning of
+        10 means samples are returned every 10 steps (default 100).
+    nproj : int > 0, optional
+        How often to reproject the sampling point into the feasibility
+        space. Avoids numerical issues at the cost of lower sampling. If
+        you observe many equality constraint violations with
+        `sampler.validate` you should lower this number (default None).
+    rounding : bool, optional
+        Whether to precondition the sampling problem by applying an offline rounding
+        transformation.
+    seed : int > 0, optional
+        Sets the random number seed. Initialized to the current time stamp
+        if None (default None).
+
+    Attributes
+    ----------
+    n_samples : int
+        The total number of samples that have been generated by this
+        sampler instance.
+    problem : typing.NamedTuple
+        A NamedTuple whose attributes define the entire sampling problem in
+        matrix form.
+    fwd_idx : numpy.array
+        A numpy array having one entry for each reaction in the model,
+        containing the index of the respective forward variable.
+    rev_idx : numpy.array
+        A numpy array having one entry for each reaction in the model,
+        containing the index of the respective reverse variable.
+    sampler :
+        The hopsy sampler type.
+    mcs : list
+        List of length `processes` of hopsy.MarkovChain objects.
+    rngs : list
+        List of length `processes` of hopsy.RandomNumberGenerator objects.
+
+    """
+
+    def __init__(
+        self,
+        model: "Model",
+        sampler: hopsy.Proposal = hopsy.UniformCoordinateHitAndRunProposal,
+        thinning: int = 100,
+        processes: Optional[int] = None,
+        nproj: Optional[int] = None,
+        seed: Optional[int] = None,
+        rounding: bool = True,
+        **kwargs,
+    ) -> None:
+        """Initialize a new HopsySampler."""
+        super().__init__(model, thinning, nproj=nproj, seed=seed, **kwargs)
+
+        if self.problem.inequalities.shape[0] > 0:
+            A = self.problem.inequalities
+            lb, ub = self.problem.bounds
+            A, b = np.concat([-A, A], axis=0), np.concat([-lb, ub], axis=0)
+            A, b = A[~np.isinf(b)], b[~np.isinf(b)]
+        else:
+            # empty dummy constraint to set problem dimension
+            A = np.ones((0, self.problem.equalities.shape[1]))
+            b = np.ones(0)
+
+        problem = hopsy.Problem(A, b)
+        problem = hopsy.add_box_constraints(
+            problem,
+            self.problem.variable_bounds[0],
+            self.problem.variable_bounds[1],
+            simplify=False,
+        )
+        problem = hopsy.add_equality_constraints(
+            problem, self.problem.equalities, self.problem.b
+        )
+        problem = hopsy.round(problem, simplify=False) if rounding else problem
+
+        self.sampler = sampler
+
+        # batched transformation to better exploit vectorization
+        self.transformation = problem.transformation.copy()
+        self.shift = problem.shift.copy()
+
+        problem.transformation = None
+        problem.shift = None
+
+        self._problem = problem
+
+        if processes is None:
+            self.processes = configuration.processes
+        else:
+            self.processes = processes
+
+        self.mcs = None
+        self.rngs = None
+
+        self.mcs = [
+            hopsy.MarkovChain(self._problem, self.sampler)
+            for i in range(self.processes)
+        ]
+        self.rngs = [
+            hopsy.RandomNumberGenerator(self._seed, i) for i in range(self.processes)
+        ]
+
+    def sample(self, n: int, fluxes: bool = True) -> pd.DataFrame:
+        """Generate a set of samples by calling hopsy.
 
         Parameters
         ----------
-        problem : hopsy.Problem
-          A hopsy.Problem with populated `problem.transformation` and `problem.shift` of
-          shape `(m, d)` and `(d, )`, respectively. If not populated, the identity
-          operation is performed.
-        samples : np.array
-          Parameter samples of shape `(..., d)` which are to be transformed.
+        n : int
+            The minimum number of samples that are generated at once.
+        fluxes : bool, optional
+            Whether to return fluxes or the internal solver variables. If
+            set to False, will return a variable for each forward and
+            backward flux as well as all additional variables you might
+            have defined in the model (default True).
 
         Returns
         -------
-        np.array
-          Transformed samples of shape `(..., m)`
+        pandas.DataFrame
+            Returns a pandas DataFrame with `n` rows, each containing a
+            flux sample.
         """
-        S, h = problem.transformation, problem.shift
-        transformed = samples @ S.T if S is not None else samples
-        shifted = transformed + h if h is not None else transformed
-        return shifted
+        _, samples = hopsy.sample(
+            self.mcs,
+            self.rngs,
+            n_samples=int(np.ceil(n / self.processes)),
+            thinning=self.thinning,
+            n_procs=self.processes,
+        )
 
-    class HopsySampler(HRSampler):
-        """
-        Generic sampling wrapper for hopsy samplers.
+        # batched transformation to better exploit vectorization
+        self._problem.transformation = self.transformation
+        self._problem.shift = self.shift
 
-        Parameters
-        ----------
-        model : cobra.Model
-            The cobra model from which to generate samples.
-        sampler : hopsy.Proposal, optional
-            The hopsy sampler for sampling the polytope
-            (default hopsy.UniformCoordinateHitAndRunProposal).
-        processes: int, optional
-            The number of processes used during sampling
-            (default cobra.Configuration.processes).
-        thinning : int, optional
-            The thinning factor of the generated sampling chain. A thinning of
-            10 means samples are returned every 10 steps (default 100).
-        nproj : int > 0, optional
-            How often to reproject the sampling point into the feasibility
-            space. Avoids numerical issues at the cost of lower sampling. If
-            you observe many equality constraint violations with
-            `sampler.validate` you should lower this number (default None).
-        rounding : bool, optional
-            Whether to precondition the sampling problem by applying an offline rounding
-            transformation.
-        seed : int > 0, optional
-            Sets the random number seed. Initialized to the current time stamp
-            if None (default None).
+        samples = back_transform(self._problem, samples)
 
-        Attributes
-        ----------
-        n_samples : int
-            The total number of samples that have been generated by this
-            sampler instance.
-        problem : typing.NamedTuple
-            A NamedTuple whose attributes define the entire sampling problem in
-            matrix form.
-        fwd_idx : numpy.array
-            A numpy array having one entry for each reaction in the model,
-            containing the index of the respective forward variable.
-        rev_idx : numpy.array
-            A numpy array having one entry for each reaction in the model,
-            containing the index of the respective reverse variable.
-        sampler :
-            The hopsy sampler type.
-        mcs : list
-            List of length `processes` of hopsy.MarkovChain objects.
-        rngs : list
-            List of length `processes` of hopsy.RandomNumberGenerator objects.
+        self._problem.transformation = None
+        self._problem.shift = None
 
-        """
+        samples = samples.reshape(-1, samples.shape[-1])  # flatten chain dim
 
-        def __init__(
-            self,
-            model: "Model",
-            sampler: hopsy.Proposal = hopsy.UniformCoordinateHitAndRunProposal,
-            thinning: int = 100,
-            processes: Optional[int] = None,
-            nproj: Optional[int] = None,
-            seed: Optional[int] = None,
-            rounding: bool = True,
-            **kwargs,
-        ) -> None:
-            """Initialize a new HopsySampler."""
-            super().__init__(model, thinning, nproj=nproj, seed=seed, **kwargs)
+        if fluxes:
+            names = [r.id for r in self.model.reactions]
 
-            if self.problem.inequalities.shape[0] > 0:
-                A = self.problem.inequalities
-                lb, ub = self.problem.bounds
-                A, b = np.concat([-A, A], axis=0), np.concat([-lb, ub], axis=0)
-                A, b = A[~np.isinf(b)], b[~np.isinf(b)]
-            else:
-                # empty dummy constraint to set problem dimension
-                A = np.ones((0, self.problem.equalities.shape[1]))
-                b = np.ones(0)
-
-            problem = hopsy.Problem(A, b)
-            problem = hopsy.add_box_constraints(
-                problem,
-                self.problem.variable_bounds[0],
-                self.problem.variable_bounds[1],
-                simplify=False,
+            return pd.DataFrame(
+                samples[:, self.fwd_idx] - samples[:, self.rev_idx],
+                columns=names,
             )
-            problem = hopsy.add_equality_constraints(
-                problem, self.problem.equalities, self.problem.b
-            )
-            problem = hopsy.round(problem, simplify=False) if rounding else problem
+        else:
+            names = [v.name for v in self.model.variables]
 
-            self.sampler = sampler
-
-            # batched transformation to better exploit vectorization
-            self.transformation = problem.transformation.copy()
-            self.shift = problem.shift.copy()
-
-            problem.transformation = None
-            problem.shift = None
-
-            self._problem = problem
-
-            if processes is None:
-                self.processes = configuration.processes
-            else:
-                self.processes = processes
-
-            self.mcs = None
-            self.rngs = None
-
-            self.mcs = [
-                hopsy.MarkovChain(self._problem, self.sampler)
-                for i in range(self.processes)
-            ]
-            self.rngs = [
-                hopsy.RandomNumberGenerator(self._seed, i)
-                for i in range(self.processes)
-            ]
-
-        def sample(self, n: int, fluxes: bool = True) -> pd.DataFrame:
-            """Generate a set of samples by calling hopsy.
-
-            Parameters
-            ----------
-            n : int
-                The minimum number of samples that are generated at once.
-            fluxes : bool, optional
-                Whether to return fluxes or the internal solver variables. If
-                set to False, will return a variable for each forward and
-                backward flux as well as all additional variables you might
-                have defined in the model (default True).
-
-            Returns
-            -------
-            pandas.DataFrame
-                Returns a pandas DataFrame with `n` rows, each containing a
-                flux sample.
-            """
-            _, samples = hopsy.sample(
-                self.mcs,
-                self.rngs,
-                n_samples=int(np.ceil(n / self.processes)),
-                thinning=self.thinning,
-                n_procs=self.processes,
-            )
-
-            # batched transformation to better exploit vectorization
-            self._problem.transformation = self.transformation
-            self._problem.shift = self.shift
-
-            samples = back_transform(self._problem, samples)
-
-            self._problem.transformation = None
-            self._problem.shift = None
-
-            samples = samples.reshape(-1, samples.shape[-1])  # flatten chain dim
-
-            if fluxes:
-                names = [r.id for r in self.model.reactions]
-
-                return pd.DataFrame(
-                    samples[:, self.fwd_idx] - samples[:, self.rev_idx],
-                    columns=names,
-                )
-            else:
-                names = [v.name for v in self.model.variables]
-
-                return pd.DataFrame(samples, columns=names)
-
-except ModuleNotFoundError:
-    hopsy_is_available = False
+            return pd.DataFrame(samples, columns=names)
